@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,60 +27,60 @@ dependencies:
 
   # File manager .
   # path: 1.6.2
-  path_provider: ^1.6.27
-  flutter_file_manager: ^0.2.0
-  id3: ^0.1.8
-  dart_tags: ^0.3.1
+  path_provider: #^1.6.27
+  flutter_file_manager: #^0.2.0
+  id3: #^0.1.8
+  dart_tags: #^0.3.1
 
-  path: ^1.6.4
-  path_provider_ex: ^1.0.1
+  path: #^1.6.4
+  path_provider_ex: #^1.0.1
 
-  english_words: ^3.1.5
-  fluttertoast: ^7.1.6
-  flutter_svg: ^0.19.2+1
+  english_words: #^3.1.5
+  fluttertoast: #^7.1.6
+  flutter_svg: #^0.19.2+1
   percent_indicator: "^2.1.7+2"
-  media_metadata_plugin: ^0.0.5
-  shimmer: ^1.1.2
-  http: ^0.12.2
-  audio_service: ^0.16.1
-  just_audio: ^0.6.5
-  flutter_tts: ^2.0.0
-  shared_preferences: ^0.5.12+4
-  lottie: ^0.7.0+1
-  carousel_slider: ^2.3.1
-  palette_generator: ^0.2.3
-  contained_tab_bar_view: ^0.7.1
-  bottom_sheet_bar: ^1.0.2+7
-  equalizer: ^0.0.2+2
-  flutter_xlider: ^3.4.0
+  media_metadata_plugin: #^0.0.5
+  shimmer: #^1.1.2
+  http: #^0.12.2
+  audio_service: #^0.16.1
+  just_audio: #^0.6.5
+  flutter_tts: #^2.0.0
+  shared_preferences: #^0.5.12+4
+  lottie: #^0.7.0+1
+  carousel_slider: #^2.3.1
+  palette_generator: #^0.2.3
+  contained_tab_bar_view: #^0.7.1
+  bottom_sheet_bar: #^1.0.2+7
+  equalizer: #^0.0.2+2
+  flutter_xlider: #^3.4.0
   #audioplayers: ^0.17.2
   #simple_permissions: ^0.1.9
   #audio_picker: ^1.0.0
-  f_logs: ^1.2.2
-  logger: ^0.9.4
-  media_store: ^1.0.0
-  video_thumbnail: ^0.2.5+1
-  audiotagger: ^1.1.0
-  sqflite: ^1.3.2+2
-  permission_handler: ^5.0.1+1
-  permission: ^0.1.7
-  provider: ^4.3.3
-  flutter_visualizers: ^0.0.2
-  wave_generator: ^0.1.0
+  f_logs: #^1.2.2
+  logger: #^0.9.4
+  media_store: #^1.0.0
+  video_thumbnail: #^0.2.5+1
+  audiotagger: #^1.1.0
+  sqflite: #^1.3.2+2
+  permission_handler: #^5.0.1+1
+  permission: #^0.1.7
+  provider: #^4.3.3
+  flutter_visualizers: #^0.0.2
+  wave_generator: #^0.1.0
   #flutter_exoplayer: ^0.5.5
-  bass_boost: ^0.1.1
-  volume: ^1.0.1
-  modal_bottom_sheet: ^1.0.0+1
-  rxdart: ^0.25.0
-  get_it: ^5.0.6
-  injectable: ^1.0.7
-  retrofit: ^1.3.4+1
-  json_annotation: ^3.1.1
+  bass_boost: #^0.1.1
+  volume: #^1.0.1
+  modal_bottom_sheet: #^1.0.0+1
+  rxdart: #^0.25.0
+  get_it: #^5.0.6
+  injectable: #^1.0.7
+  retrofit: #^1.3.4+1
+  json_annotation: #^3.1.1
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.1
+  cupertino_icons: #^1.0.1
 
 
 dev_dependencies:
