@@ -1,8 +1,8 @@
 class ItemModel {
-  int _page;
-  int _total_results;
-  int _total_pages;
-  List<_Result> _results = [];
+  int? _page;
+  int? _total_results;
+  int? _total_pages;
+  List<_Result>? _results = [];
 
   ItemModel.fromJson(Map<String, dynamic> parsedJson) {
     print(parsedJson['results'].length);
@@ -17,30 +17,30 @@ class ItemModel {
     _results = temp;
   }
 
-  List<_Result> get results => _results;
+  List<_Result>? get results => _results;
 
-  int get total_pages => _total_pages;
+  int? get total_pages => _total_pages;
 
-  int get total_results => _total_results;
+  int? get total_results => _total_results;
 
-  int get page => _page;
+  int? get page => _page;
 }
 
 class _Result {
-  int _vote_count;
-  int _id;
-  bool _video;
+  int? _vote_count;
+  int? _id;
+  bool? _video;
   var _vote_average;
-  String _title;
-  double _popularity;
-  String _poster_path;
-  String _original_language;
-  String _original_title;
-  List<int> _genre_ids = [];
-  String _backdrop_path;
-  bool _adult;
-  String _overview;
-  String _release_date;
+  String? _title;
+  double? _popularity;
+  String? _poster_path;
+  String? _original_language;
+  String? _original_title;
+  List<int>? _genre_ids = [];
+  String? _backdrop_path;
+  bool? _adult;
+  String? _overview;
+  String? _release_date;
 
   _Result(result) {
     _vote_count = result['vote_count'];
@@ -53,7 +53,7 @@ class _Result {
     _original_language = result['original_language'];
     _original_title = result['original_title'];
     for (int i = 0; i < result['genre_ids'].length; i++) {
-      _genre_ids.add(result['genre_ids'][i]);
+      _genre_ids?.add(result['genre_ids'][i]);
     }
     _backdrop_path = result['backdrop_path'];
     _adult = result['adult'];
@@ -61,31 +61,31 @@ class _Result {
     _release_date = result['release_date'];
   }
 
-  String get release_date => _release_date;
+  String? get release_date => _release_date;
 
-  String get overview => _overview;
+  String? get overview => _overview;
 
-  bool get adult => _adult;
+  bool? get adult => _adult;
 
-  String get backdrop_path => _backdrop_path;
+  String? get backdrop_path => _backdrop_path;
 
-  List<int> get genre_ids => _genre_ids;
+  List<int>? get genre_ids => _genre_ids;
 
-  String get original_title => _original_title;
+  String? get original_title => _original_title;
 
-  String get original_language => _original_language;
+  String? get original_language => _original_language;
 
-  String get poster_path => _poster_path;
+  String? get poster_path => _poster_path;
 
-  double get popularity => _popularity;
+  double? get popularity => _popularity;
 
-  String get title => _title;
+  String? get title => _title;
 
   get vote_average => _vote_average;
 
-  bool get video => _video;
+  bool? get video => _video;
 
-  int get id => _id;
+  int? get id => _id;
 
-  int get vote_count => _vote_count;
+  int? get vote_count => _vote_count;
 }
