@@ -2,8 +2,8 @@ import 'package:flutter/cupertino.dart';
 
 class HomeView extends StatelessWidget {
 // Home View has a dependency on the AppInfo
-  AppInfo appInfo;
-  HomeView({Key key}) : super(key: key);
+  AppInfo? appInfo;
+  HomeView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
