@@ -9,9 +9,9 @@ void main() {
 }
 
 class BottomSheetBarPage extends StatefulWidget {
-  final String title;
+  final String? title;
 
-  BottomSheetBarPage({Key key, this.title}) : super(key: key);
+  BottomSheetBarPage({Key? key, this.title}) : super(key: key);
 
   @override
   _BottomSheetBarPageState createState() => _BottomSheetBarPageState();
@@ -102,10 +102,12 @@ class _BottomSheetBarPageState extends State<BottomSheetBarPage> {
             topLeft: Radius.circular(0.0),
             topRight: Radius.circular(0.0),
           ),
-          expandedBuilder: (scrollController) => Expand(),
+          //Todo : thanh comment ...
+          expandedBuilder: (scrollController) => Container(),
           body: Container(
             color: Colors.blue,
           ),
+          //Expand(),
         ),
       );
 }
