@@ -221,9 +221,9 @@ final String columnFrequency = 'frequency';
 
 // data model class
 class Word {
-  int id;
-  String word;
-  int frequency;
+  int? id;
+  String? word;
+  int? frequency;
 
   Word();
 
@@ -258,12 +258,12 @@ class DatabaseHelper {
   static final DatabaseHelper instance = DatabaseHelper._privateConstructor();
 
   // Only allow a single open connection to the database.
-  static Database _database;
+  static Database? _database;
 
   Future<Database> get database async {
-    if (_database != null) return _database;
+    if (_database != null) return _database!;
     _database = await _initDatabase();
-    return _database;
+    return _database!;
   }
 
   // open the database
@@ -294,14 +294,14 @@ class DatabaseHelper {
     return id;
   }
 
-  Future<Word> queryWord(int id) async {
+  Future<Word?> queryWord(int id) async {
     Database db = await database;
     List<Map> maps = await db.query(tableWords,
         columns: [columnId, columnWord, columnFrequency],
         where: '$columnId = ?',
         whereArgs: [id]);
     if (maps.length > 0) {
-      return Word.fromMap(maps.first);
+      return Word.fromMap(maps.first as Map<String, dynamic>);
     }
     return null;
   }
