@@ -10,26 +10,26 @@ class ProgressIndicatorDemo extends StatefulWidget {
 
 class _ProgressIndicatorDemoState extends State<ProgressIndicatorDemo>
     with SingleTickerProviderStateMixin {
-  AnimationController controller;
-  Animation<double> animation;
+  AnimationController? controller;
+  Animation<double>? animation;
 
   @override
   void initState() {
     super.initState();
     controller = AnimationController(
         duration: const Duration(milliseconds: 2000), vsync: this);
-    animation = Tween(begin: 0.0, end: 1.0).animate(controller)
+    animation = Tween(begin: 0.0, end: 1.0).animate(controller!)
       ..addListener(() {
         setState(() {
           // the state that has changed here is the animation object’s value
         });
       });
-    controller.repeat();
+    controller?.repeat();
   }
 
   @override
   void dispose() {
-    controller.stop();
+    controller?.stop();
     super.dispose();
   }
 
@@ -38,7 +38,7 @@ class _ProgressIndicatorDemoState extends State<ProgressIndicatorDemo>
     return new Center(
         child: new Container(
           child: LinearProgressIndicator(
-            value: animation.value,
+            value: animation!.value,
           ),
         ));
   }
