@@ -4,7 +4,7 @@ import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:audio_service/audio_service.dart';
-import 'package:audio_session/audio_session.dart';
+//import 'package:audio_session/audio_session.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_app/example/main/app/data/DatabaseUtils.dart';
@@ -36,54 +36,56 @@ class MainScreen extends StatelessWidget {
     " AudioService.start :.. ".Log();
     AudioService.stop();
     AudioService.start(
-      backgroundTaskEntrypoint: _audioPlayerTaskEntrypoint,
+      //backgroundTaskEntrypoint: _audioPlayerTaskEntrypoint,
       androidNotificationChannelName: 'Audio Service Demo',
       // Enable this if you want the Android service to exit the foreground state on pause.
       // androidStopForegroundOnPause: true,
-      androidNotificationColor: 0xFF2196f3,
-      androidNotificationIcon: 'mipmap/ic_launcher',
-      androidEnableQueue: true,
+      //androidNotificationColor: 0xFF2196f3,
+      //androidNotificationIcon: 'mipmap/ic_launcher',
+      //androidEnableQueue: true,
     );
 
-    return AudioServiceWidget(
-        child: Container(
-      child: Center(
-        child: StreamBuilder<bool>(
-          stream: AudioService.runningStream,
-          builder: (context, snapshot) {
-            return Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                // Play/pause/stop buttons.
-                StreamBuilder<bool>(
-                  stream: AudioService.playbackStateStream
-                      .map((state) => state.playing)
-                      .distinct(),
-                  builder: (context, snapshot) {
-                    final playing = snapshot.data ?? false;
-                    return Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        if (playing) pauseButton() else playButton(),
-                      ],
-                    );
-                  },
-                ),
-              ],
-            );
-          },
-        ),
-      ),
-    ));
+    // return AudioServiceWidget(
+    //     child: Container(
+    //   child: Center(
+    //     child: StreamBuilder<bool>(
+    //       stream: AudioService.runningStream,
+    //       builder: (context, snapshot) {
+    //         return Column(
+    //           mainAxisAlignment: MainAxisAlignment.center,
+    //           children: [
+    //             // Play/pause/stop buttons.
+    //             StreamBuilder<bool>(
+    //               stream: AudioService.playbackStateStream
+    //                   .map((state) => state.playing)
+    //                   .distinct(),
+    //               builder: (context, snapshot) {
+    //                 final playing = snapshot.data ?? false;
+    //                 return Row(
+    //                   mainAxisAlignment: MainAxisAlignment.center,
+    //                   children: [
+    //                     if (playing) pauseButton() else playButton(),
+    //                   ],
+    //                 );
+    //               },
+    //             ),
+    //           ],
+    //         );
+    //       },
+    //     ),
+    //   ),
+    // ));
+
+    return Container();
   }
 
   RaisedButton textToSpeechButton() => startButton(
         'TextToSpeech',
         () {
           AudioService.start(
-            backgroundTaskEntrypoint: _textToSpeechTaskEntrypoint,
+            //backgroundTaskEntrypoint: _textToSpeechTaskEntrypoint,
             androidNotificationChannelName: 'Audio Service Demo',
-            androidNotificationColor: 0xFF2196f3,
+            //androidNotificationColor: 0xFF2196f3,
             androidNotificationIcon: 'mipmap/ic_launcher',
           );
         },
@@ -128,310 +130,309 @@ class MediaState {
   MediaState(this.mediaItem, this.position);
 }
 
-class SeekBar extends StatefulWidget {
-  final Duration duration;
-  final Duration position;
-  final ValueChanged<Duration> onChanged;
-  final ValueChanged<Duration> onChangeEnd;
+// class SeekBar extends StatefulWidget {
+//   final Duration duration;
+//   final Duration position;
+//   final ValueChanged<Duration> onChanged;
+//   final ValueChanged<Duration> onChangeEnd;
+//
+//   SeekBar({
+//     @required this.duration,
+//     @required this.position,
+//     this.onChanged,
+//     this.onChangeEnd,
+//   });
+//
+//   @override
+//   _SeekBarState createState() => _SeekBarState();
+// }
 
-  SeekBar({
-    @required this.duration,
-    @required this.position,
-    this.onChanged,
-    this.onChangeEnd,
-  });
-
-  @override
-  _SeekBarState createState() => _SeekBarState();
-}
-
-class _SeekBarState extends State<SeekBar> {
-  double _dragValue;
-  bool _dragging = false;
-
-  @override
-  Widget build(BuildContext context) {
-    final value = min(_dragValue ?? widget.position?.inMilliseconds?.toDouble(),
-        widget.duration.inMilliseconds.toDouble());
-    if (_dragValue != null && !_dragging) {
-      _dragValue = null;
-    }
-    return Stack(
-      children: [
-        Slider(
-          min: 0.0,
-          max: widget.duration.inMilliseconds.toDouble(),
-          value: value,
-          onChanged: (value) {
-            if (!_dragging) {
-              _dragging = true;
-            }
-            setState(() {
-              _dragValue = value;
-            });
-            if (widget.onChanged != null) {
-              widget.onChanged(Duration(milliseconds: value.round()));
-            }
-          },
-          onChangeEnd: (value) {
-            if (widget.onChangeEnd != null) {
-              widget.onChangeEnd(Duration(milliseconds: value.round()));
-            }
-            _dragging = false;
-          },
-        ),
-        Positioned(
-          right: 16.0,
-          bottom: 0.0,
-          child: Text(
-              RegExp(r'((^0*[1-9]\d*:)?\d{2}:\d{2})\.\d+$')
-                      .firstMatch("$_remaining")
-                      ?.group(1) ??
-                  '$_remaining',
-              style: Theme.of(context).textTheme.caption),
-        ),
-      ],
-    );
-  }
-
-  Duration get _remaining => widget.duration - widget.position;
-}
+// class _SeekBarState extends State<SeekBar> {
+//   double? _dragValue;
+//   bool? _dragging = false;
+//
+//   @override
+//   Widget build(BuildContext context) {
+//     final value = min(_dragValue ?? widget.position.inMilliseconds.toDouble(), widget.duration.inMilliseconds.toDouble());
+//     if (_dragValue != null && !_dragging!) {
+//       _dragValue = null;
+//     }
+//     return Stack(
+//       children: [
+//         Slider(
+//           min: 0.0,
+//           max: widget.duration.inMilliseconds.toDouble(),
+//           value: value,
+//           onChanged: (value) {
+//             if (!_dragging!) {
+//               _dragging = true;
+//             }
+//             setState(() {
+//               _dragValue = value;
+//             });
+//             if (widget.onChanged != null) {
+//               widget.onChanged(Duration(milliseconds: value.round()));
+//             }
+//           },
+//           onChangeEnd: (value) {
+//             if (widget.onChangeEnd != null) {
+//               widget.onChangeEnd(Duration(milliseconds: value.round()));
+//             }
+//             _dragging = false;
+//           },
+//         ),
+//         Positioned(
+//           right: 16.0,
+//           bottom: 0.0,
+//           child: Text(
+//               RegExp(r'((^0*[1-9]\d*:)?\d{2}:\d{2})\.\d+$')
+//                       .firstMatch("$_remaining")
+//                       ?.group(1) ??
+//                   '$_remaining',
+//               style: Theme.of(context).textTheme.caption),
+//         ),
+//       ],
+//     );
+//   }
+//
+//   Duration get _remaining => widget.duration - widget.position;
+// }
 
 // NOTE: Your entrypoint MUST be a top-level function.
 void _audioPlayerTaskEntrypoint() async {
-  AudioServiceBackground.run(() => AudioPlayerTask());
+  //AudioServiceBackground.run(() => AudioPlayerTask());
 }
 
 /// This task defines logic for playing a list of podcast episodes.
-class AudioPlayerTask extends BackgroundAudioTask {
-  final _mediaLibrary = MediaLibrary();
-  AudioPlayer _player = new AudioPlayer();
-  AudioProcessingState _skipState;
-  Seeker _seeker;
-  StreamSubscription<PlaybackEvent> _eventSubscription;
-
-  //List<MediaItem> get queue => _mediaLibrary.items;
-  List<MusicModel> queue;
-
-  int get index => _player.currentIndex;
-
-  _getMediaItem(MusicModel musicModel) {
-    var mediaItem = MediaItem(
-      id: musicModel.url,
-      album: musicModel.album,
-      title: musicModel.artist,
-      artist: musicModel.artist,
-      duration: Duration(milliseconds: musicModel.duration),
-      artUri:
-          "https://image.freepik.com/free-photo/white-clean-brick-wall-as-texture-background-backdrop-high-resolution-picture_117930-181.jpg", //Uint8List.fromList(musicModel.logoMemory.codeUnits),//"https://media.wnyc.org/i/1400/1400/l/80/1/ScienceFriday_WNYCStudios_1400.jpg",
-    );
-    return mediaItem;
-  }
-
-  _getListMediaItem(List<MusicModel> data) {
-    List<MediaItem> listMedialItem = [];
-    data.forEach((element) {
-      listMedialItem.add(_getMediaItem(element));
-    });
-    return listMedialItem;
-  }
-
-  MediaItem get mediaItem => index == null ? null : _getMediaItem(queue[index]);
-
-  @override
-  Future<void> onStart(Map<String, dynamic> params) async {
-    "onStart :... ".LogP();
-
-    handle() async {
-      "handle :...".Log();
-      // We configure the audio session for speech since we're playing a podcast.
-      // You can also put this in your app's initialisation if your app doesn't
-      // switch between two types of audio as this example does.
-      final session = await AudioSession.instance;
-      await session.configure(AudioSessionConfiguration.speech());
-      // Broadcast media item changes.
-      _player.currentIndexStream.listen((index) {
-        var musicData = queue[index];
-        var mediaItem = _getMediaItem(musicData);
-        if (index != null) AudioServiceBackground.setMediaItem(mediaItem);
-      });
-
-      // Propagate all events from the audio player to AudioService clients.
-      _eventSubscription = _player.playbackEventStream.listen((event) {
-        _broadcastState();
-      });
-      // Special processing for state transitions.
-      _player.processingStateStream.listen((state) {
-        switch (state) {
-          case ProcessingState.completed:
-            // In this example, the service stops when reaching the end.
-            onStop();
-            break;
-          case ProcessingState.ready:
-            // If we just came from skipping between tracks, clear the skip
-            // state now that we're ready to play.
-            _skipState = null;
-            break;
-          default:
-            break;
-        }
-      });
-
-      //Todo :  /storage/emulated/0/NCT/DiDuDuaDi-BichPhuong-6059493.mp3
-
-      // Load and broadcast the queue
-      AudioServiceBackground.setQueue(_getListMediaItem(queue));
-      try {
-        await _player.setAudioSource(ConcatenatingAudioSource(
-          children:
-              queue.map((item) => AudioSource.uri(Uri.file(item.url))).toList(),
-        ));
-        // In this example, we automatically start playing on start.
-        onPlay();
-      } catch (e) {
-        print("Error: $e");
-        onStop();
-      }
-    }
-
-    var databaseUtils = DatabaseUtils.instance;
-    databaseUtils.getAll().then((data) {
-      "data :.. ${data.length} ".Log();
-      queue = data;
-
-      //Todo : start :...
-      handle();
-      //Todo :end :...
-    });
-  }
-
-  @override
-  Future<void> onSkipToQueueItem(String mediaId) async {
-    //"mediaId :... ${mediaId} ".Log();
-    //"onSkipToQueueItem :.. ${queue.length} } ".Log();
-
-    // Then default implementations of onSkipToNext and onSkipToPrevious will
-    // delegate to this method.
-    final newIndex = queue.indexWhere((MusicModel item) => item.url == mediaId);
-    if (newIndex == -1) return;
-    // During a skip, the player may enter the buffering state. We could just
-    // propagate that state directly to AudioService clients but AudioService
-    // has some more specific states we could use for skipping to next and
-    // previous. This variable holds the preferred state to send instead of
-    // buffering during a skip, and it is cleared as soon as the player exits
-    // buffering (see the listener in onStart).
-    _skipState = newIndex > index
-        ? AudioProcessingState.skippingToNext
-        : AudioProcessingState.skippingToPrevious;
-    // This jumps to the beginning of the queue item at newIndex.
-    _player.seek(Duration.zero, index: newIndex);
-  }
-
-  playLocal() async {
-    // AudioPlayer audioPlayer = AudioPlayer();
-    //int result = await audioPlayer.play(path, isLocal: true);
-  }
-
-  @override
-  Future<void> onPlay() => _player.play(); //_player.play();
-
-  @override
-  Future<void> onPause() => _player.pause();
-
-  @override
-  Future<void> onSeekTo(Duration position) => _player.seek(position);
-
-  @override
-  Future<void> onFastForward() => _seekRelative(fastForwardInterval);
-
-  @override
-  Future<void> onRewind() => _seekRelative(-rewindInterval);
-
-  @override
-  Future<void> onSeekForward(bool begin) async => _seekContinuously(begin, 1);
-
-  @override
-  Future<void> onSeekBackward(bool begin) async => _seekContinuously(begin, -1);
-
-  @override
-  Future<void> onStop() async {
-    "stop :.. ".Log();
-
-    await _player.dispose();
-    _eventSubscription.cancel();
-    // It is important to wait for this state to be broadcast before we shut
-    // down the task. If we don't, the background task will be destroyed before
-    // the message gets sent to the UI.
-    await _broadcastState();
-    // Shut down this task
-    await super.onStop();
-  }
-
-  /// Jumps away from the current position by [offset].
-  Future<void> _seekRelative(Duration offset) async {
-    var newPosition = _player.position + offset;
-    // Make sure we don't jump out of bounds.
-    if (newPosition < Duration.zero) newPosition = Duration.zero;
-    if (newPosition > mediaItem.duration) newPosition = mediaItem.duration;
-    // Perform the jump via a seek.
-    await _player.seek(newPosition);
-  }
-
-  /// Begins or stops a continuous seek in [direction]. After it begins it will
-  /// continue seeking forward or backward by 10 seconds within the audio, at
-  /// intervals of 1 second in app time.
-  void _seekContinuously(bool begin, int direction) {
-    _seeker?.stop();
-    if (begin) {
-      _seeker = Seeker(_player, Duration(seconds: 10 * direction),
-          Duration(seconds: 1), mediaItem)
-        ..start();
-    }
-  }
-
-  /// Broadcasts the current state to all clients.
-  Future<void> _broadcastState() async {
-    await AudioServiceBackground.setState(
-      controls: [
-        MediaControl.skipToPrevious,
-        if (_player.playing) MediaControl.pause else MediaControl.play,
-        MediaControl.stop,
-        MediaControl.skipToNext,
-      ],
-      systemActions: [
-        MediaAction.seekTo,
-        MediaAction.seekForward,
-        MediaAction.seekBackward,
-      ],
-      androidCompactActions: [0, 1, 3],
-      processingState: _getProcessingState(),
-      playing: _player.playing,
-      position: _player.position,
-      bufferedPosition: _player.bufferedPosition,
-      speed: _player.speed,
-    );
-  }
-
-  /// Maps just_audio's processing state into into audio_service's playing
-  /// state. If we are in the middle of a skip, we use [_skipState] instead.
-  AudioProcessingState _getProcessingState() {
-    if (_skipState != null) return _skipState;
-    switch (_player.processingState) {
-      case ProcessingState.idle:
-        return AudioProcessingState.stopped;
-      case ProcessingState.loading:
-        return AudioProcessingState.connecting;
-      case ProcessingState.buffering:
-        return AudioProcessingState.buffering;
-      case ProcessingState.ready:
-        return AudioProcessingState.ready;
-      case ProcessingState.completed:
-        return AudioProcessingState.completed;
-      default:
-        throw Exception("Invalid state: ${_player.processingState}");
-    }
-  }
-}
+// class AudioPlayerTask extends BackgroundAudioTask {
+//   final _mediaLibrary = MediaLibrary();
+//   AudioPlayer _player = new AudioPlayer();
+//   AudioProcessingState _skipState;
+//   Seeker _seeker;
+//   StreamSubscription<PlaybackEvent> _eventSubscription;
+//
+//   //List<MediaItem> get queue => _mediaLibrary.items;
+//   List<MusicModel> queue;
+//
+//   int get index => _player.currentIndex;
+//
+//   _getMediaItem(MusicModel musicModel) {
+//     var mediaItem = MediaItem(
+//       id: musicModel.url,
+//       album: musicModel.album,
+//       title: musicModel.artist,
+//       artist: musicModel.artist,
+//       duration: Duration(milliseconds: musicModel.duration),
+//       artUri:
+//           "https://image.freepik.com/free-photo/white-clean-brick-wall-as-texture-background-backdrop-high-resolution-picture_117930-181.jpg", //Uint8List.fromList(musicModel.logoMemory.codeUnits),//"https://media.wnyc.org/i/1400/1400/l/80/1/ScienceFriday_WNYCStudios_1400.jpg",
+//     );
+//     return mediaItem;
+//   }
+//
+//   _getListMediaItem(List<MusicModel> data) {
+//     List<MediaItem> listMedialItem = [];
+//     data.forEach((element) {
+//       listMedialItem.add(_getMediaItem(element));
+//     });
+//     return listMedialItem;
+//   }
+//
+//   MediaItem get mediaItem => index == null ? null : _getMediaItem(queue[index]);
+//
+//   @override
+//   Future<void> onStart(Map<String, dynamic> params) async {
+//     "onStart :... ".LogP();
+//
+//     handle() async {
+//       "handle :...".Log();
+//       // We configure the audio session for speech since we're playing a podcast.
+//       // You can also put this in your app's initialisation if your app doesn't
+//       // switch between two types of audio as this example does.
+//       final session = await AudioSession.instance;
+//       await session.configure(AudioSessionConfiguration.speech());
+//       // Broadcast media item changes.
+//       _player.currentIndexStream.listen((index) {
+//         var musicData = queue[index];
+//         var mediaItem = _getMediaItem(musicData);
+//         if (index != null) AudioServiceBackground.setMediaItem(mediaItem);
+//       });
+//
+//       // Propagate all events from the audio player to AudioService clients.
+//       _eventSubscription = _player.playbackEventStream.listen((event) {
+//         _broadcastState();
+//       });
+//       // Special processing for state transitions.
+//       _player.processingStateStream.listen((state) {
+//         switch (state) {
+//           case ProcessingState.completed:
+//             // In this example, the service stops when reaching the end.
+//             onStop();
+//             break;
+//           case ProcessingState.ready:
+//             // If we just came from skipping between tracks, clear the skip
+//             // state now that we're ready to play.
+//             _skipState = null;
+//             break;
+//           default:
+//             break;
+//         }
+//       });
+//
+//       //Todo :  /storage/emulated/0/NCT/DiDuDuaDi-BichPhuong-6059493.mp3
+//
+//       // Load and broadcast the queue
+//       AudioServiceBackground.setQueue(_getListMediaItem(queue));
+//       try {
+//         await _player.setAudioSource(ConcatenatingAudioSource(
+//           children:
+//               queue.map((item) => AudioSource.uri(Uri.file(item.url))).toList(),
+//         ));
+//         // In this example, we automatically start playing on start.
+//         onPlay();
+//       } catch (e) {
+//         print("Error: $e");
+//         onStop();
+//       }
+//     }
+//
+//     var databaseUtils = DatabaseUtils.instance;
+//     databaseUtils.getAll().then((data) {
+//       "data :.. ${data.length} ".Log();
+//       queue = data;
+//
+//       //Todo : start :...
+//       handle();
+//       //Todo :end :...
+//     });
+//   }
+//
+//   @override
+//   Future<void> onSkipToQueueItem(String mediaId) async {
+//     //"mediaId :... ${mediaId} ".Log();
+//     //"onSkipToQueueItem :.. ${queue.length} } ".Log();
+//
+//     // Then default implementations of onSkipToNext and onSkipToPrevious will
+//     // delegate to this method.
+//     final newIndex = queue.indexWhere((MusicModel item) => item.url == mediaId);
+//     if (newIndex == -1) return;
+//     // During a skip, the player may enter the buffering state. We could just
+//     // propagate that state directly to AudioService clients but AudioService
+//     // has some more specific states we could use for skipping to next and
+//     // previous. This variable holds the preferred state to send instead of
+//     // buffering during a skip, and it is cleared as soon as the player exits
+//     // buffering (see the listener in onStart).
+//     _skipState = newIndex > index
+//         ? AudioProcessingState.skippingToNext
+//         : AudioProcessingState.skippingToPrevious;
+//     // This jumps to the beginning of the queue item at newIndex.
+//     _player.seek(Duration.zero, index: newIndex);
+//   }
+//
+//   playLocal() async {
+//     // AudioPlayer audioPlayer = AudioPlayer();
+//     //int result = await audioPlayer.play(path, isLocal: true);
+//   }
+//
+//   @override
+//   Future<void> onPlay() => _player.play(); //_player.play();
+//
+//   @override
+//   Future<void> onPause() => _player.pause();
+//
+//   @override
+//   Future<void> onSeekTo(Duration position) => _player.seek(position);
+//
+//   @override
+//   Future<void> onFastForward() => _seekRelative(fastForwardInterval);
+//
+//   @override
+//   Future<void> onRewind() => _seekRelative(-rewindInterval);
+//
+//   @override
+//   Future<void> onSeekForward(bool begin) async => _seekContinuously(begin, 1);
+//
+//   @override
+//   Future<void> onSeekBackward(bool begin) async => _seekContinuously(begin, -1);
+//
+//   @override
+//   Future<void> onStop() async {
+//     "stop :.. ".Log();
+//
+//     await _player.dispose();
+//     _eventSubscription.cancel();
+//     // It is important to wait for this state to be broadcast before we shut
+//     // down the task. If we don't, the background task will be destroyed before
+//     // the message gets sent to the UI.
+//     await _broadcastState();
+//     // Shut down this task
+//     await super.onStop();
+//   }
+//
+//   /// Jumps away from the current position by [offset].
+//   Future<void> _seekRelative(Duration offset) async {
+//     var newPosition = _player.position + offset;
+//     // Make sure we don't jump out of bounds.
+//     if (newPosition < Duration.zero) newPosition = Duration.zero;
+//     if (newPosition > mediaItem.duration) newPosition = mediaItem.duration;
+//     // Perform the jump via a seek.
+//     await _player.seek(newPosition);
+//   }
+//
+//   /// Begins or stops a continuous seek in [direction]. After it begins it will
+//   /// continue seeking forward or backward by 10 seconds within the audio, at
+//   /// intervals of 1 second in app time.
+//   void _seekContinuously(bool begin, int direction) {
+//     _seeker?.stop();
+//     if (begin) {
+//       _seeker = Seeker(_player, Duration(seconds: 10 * direction),
+//           Duration(seconds: 1), mediaItem)
+//         ..start();
+//     }
+//   }
+//
+//   /// Broadcasts the current state to all clients.
+//   Future<void> _broadcastState() async {
+//     await AudioServiceBackground.setState(
+//       controls: [
+//         MediaControl.skipToPrevious,
+//         if (_player.playing) MediaControl.pause else MediaControl.play,
+//         MediaControl.stop,
+//         MediaControl.skipToNext,
+//       ],
+//       systemActions: [
+//         MediaAction.seekTo,
+//         MediaAction.seekForward,
+//         MediaAction.seekBackward,
+//       ],
+//       androidCompactActions: [0, 1, 3],
+//       processingState: _getProcessingState(),
+//       playing: _player.playing,
+//       position: _player.position,
+//       bufferedPosition: _player.bufferedPosition,
+//       speed: _player.speed,
+//     );
+//   }
+//
+//   /// Maps just_audio's processing state into into audio_service's playing
+//   /// state. If we are in the middle of a skip, we use [_skipState] instead.
+//   AudioProcessingState _getProcessingState() {
+//     if (_skipState != null) return _skipState;
+//     switch (_player.processingState) {
+//       case ProcessingState.idle:
+//         return AudioProcessingState.stopped;
+//       case ProcessingState.loading:
+//         return AudioProcessingState.connecting;
+//       case ProcessingState.buffering:
+//         return AudioProcessingState.buffering;
+//       case ProcessingState.ready:
+//         return AudioProcessingState.ready;
+//       case ProcessingState.completed:
+//         return AudioProcessingState.completed;
+//       default:
+//         throw Exception("Invalid state: ${_player.processingState}");
+//     }
+//   }
+// }
 
 /// Provides access to a library of media items. In your app, this could come
 /// from a database or web service.
@@ -444,7 +445,7 @@ class MediaLibrary {
       album: "Science Friday 11",
       title: "A Salute To Head-Scratching Science 11",
       artist: "Science Friday and WNYC Studios",
-      duration: Duration(milliseconds: 57300),
+      //duration: Duration(milliseconds: 57300),
       artUri:
           "https://media.wnyc.org/i/1400/1400/l/80/1/ScienceFriday_WNYCStudios_1400.jpg",
     ),
@@ -453,7 +454,7 @@ class MediaLibrary {
       album: "Science Friday 11",
       title: "From Cat Rheology To Operatic Incompetence 11",
       artist: "Science Friday and WNYC Studios",
-      duration: Duration(milliseconds: 28500),
+      //duration: Duration(milliseconds: 28500),
       artUri:
           "https://media.wnyc.org/i/1400/1400/l/80/1/ScienceFriday_WNYCStudios_1400.jpg",
     ),
@@ -463,245 +464,245 @@ class MediaLibrary {
 }
 
 // NOTE: Your entrypoint MUST be a top-level function.
-void _textToSpeechTaskEntrypoint() async {
-  AudioServiceBackground.run(() => TextPlayerTask());
-}
+// void _textToSpeechTaskEntrypoint() async {
+//   AudioServiceBackground.run(() => TextPlayerTask());
+// }
 
 /// This task defines logic for speaking a sequence of numbers using
 /// text-to-speech.
-class TextPlayerTask extends BackgroundAudioTask {
-  Tts _tts = Tts();
-  bool _finished = false;
-  Sleeper _sleeper = Sleeper();
-  Completer _completer = Completer();
-  bool _interrupted = false;
-
-  bool get _playing => AudioServiceBackground.state.playing;
-
-  @override
-  Future<void> onStart(Map<String, dynamic> params) async {
-    // flutter_tts resets the AVAudioSession category to playAndRecord and the
-    // options to defaultToSpeaker whenever this background isolate is loaded,
-    // so we need to set our preferred audio session configuration here after
-    // that has happened.
-    final session = await AudioSession.instance;
-    await session.configure(AudioSessionConfiguration.speech());
-    // Handle audio interruptions.
-    session.interruptionEventStream.listen((event) {
-      if (event.begin) {
-        if (_playing) {
-          onPause();
-          _interrupted = true;
-        }
-      } else {
-        switch (event.type) {
-          case AudioInterruptionType.pause:
-          case AudioInterruptionType.duck:
-            if (!_playing && _interrupted) {
-              onPlay();
-            }
-            break;
-          case AudioInterruptionType.unknown:
-            break;
-        }
-        _interrupted = false;
-      }
-    });
-    // Handle unplugged headphones.
-    session.becomingNoisyEventStream.listen((_) {
-      if (_playing) onPause();
-    });
-
-    // Start playing.
-    await _playPause();
-    for (var i = 1; i <= 10 && !_finished;) {
-      AudioServiceBackground.setMediaItem(mediaItem(i));
-      AudioServiceBackground.androidForceEnableMediaButtons();
-      try {
-        await _tts.speak('$i');
-        i++;
-        await _sleeper.sleep(Duration(milliseconds: 300));
-      } catch (e) {
-        // Speech was interrupted
-      }
-      // If we were just paused
-      if (!_finished && !_playing) {
-        try {
-          // Wait to be unpaused
-          await _sleeper.sleep();
-        } catch (e) {
-          // unpaused
-        }
-      }
-    }
-    await AudioServiceBackground.setState(
-      controls: [],
-      processingState: AudioProcessingState.stopped,
-      playing: false,
-    );
-    if (!_finished) {
-      onStop();
-    }
-    _completer.complete();
-  }
-
-  @override
-  Future<void> onPlay() => _playPause();
-
-  @override
-  Future<void> onPause() => _playPause();
-
-  @override
-  Future<void> onStop() async {
-    // Signal the speech to stop
-    _finished = true;
-    _sleeper.interrupt();
-    _tts.interrupt();
-    // Wait for the speech to stop
-    await _completer.future;
-    // Shut down this task
-    await super.onStop();
-  }
-
-  MediaItem mediaItem(int number) => MediaItem(
-      id: 'tts_$number',
-      album: 'Numbers',
-      title: 'Number $number',
-      artist: 'Sample Artist');
-
-  Future<void> _playPause() async {
-    if (_playing) {
-      _interrupted = false;
-      await AudioServiceBackground.setState(
-        controls: [MediaControl.play, MediaControl.stop],
-        processingState: AudioProcessingState.ready,
-        playing: false,
-      );
-      _sleeper.interrupt();
-      _tts.interrupt();
-    } else {
-      final session = await AudioSession.instance;
-      // flutter_tts doesn't activate the session, so we do it here. This
-      // allows the app to stop other apps from playing audio while we are
-      // playing audio.
-      if (await session.setActive(true)) {
-        // If we successfully activated the session, set the state to playing
-        // and resume playback.
-        await AudioServiceBackground.setState(
-          controls: [MediaControl.pause, MediaControl.stop],
-          processingState: AudioProcessingState.ready,
-          playing: true,
-        );
-        _sleeper.interrupt();
-      }
-    }
-  }
-}
+// class TextPlayerTask extends BackgroundAudioTask {
+//   Tts _tts = Tts();
+//   bool _finished = false;
+//   Sleeper _sleeper = Sleeper();
+//   Completer _completer = Completer();
+//   bool _interrupted = false;
+//
+//   bool get _playing => AudioServiceBackground.state.playing;
+//
+//   @override
+//   Future<void> onStart(Map<String, dynamic> params) async {
+//     // flutter_tts resets the AVAudioSession category to playAndRecord and the
+//     // options to defaultToSpeaker whenever this background isolate is loaded,
+//     // so we need to set our preferred audio session configuration here after
+//     // that has happened.
+//     final session = await AudioSession.instance;
+//     await session.configure(AudioSessionConfiguration.speech());
+//     // Handle audio interruptions.
+//     session.interruptionEventStream.listen((event) {
+//       if (event.begin) {
+//         if (_playing) {
+//           onPause();
+//           _interrupted = true;
+//         }
+//       } else {
+//         switch (event.type) {
+//           case AudioInterruptionType.pause:
+//           case AudioInterruptionType.duck:
+//             if (!_playing && _interrupted) {
+//               onPlay();
+//             }
+//             break;
+//           case AudioInterruptionType.unknown:
+//             break;
+//         }
+//         _interrupted = false;
+//       }
+//     });
+//     // Handle unplugged headphones.
+//     session.becomingNoisyEventStream.listen((_) {
+//       if (_playing) onPause();
+//     });
+//
+//     // Start playing.
+//     await _playPause();
+//     for (var i = 1; i <= 10 && !_finished;) {
+//       AudioServiceBackground.setMediaItem(mediaItem(i));
+//       AudioServiceBackground.androidForceEnableMediaButtons();
+//       try {
+//         await _tts.speak('$i');
+//         i++;
+//         await _sleeper.sleep(Duration(milliseconds: 300));
+//       } catch (e) {
+//         // Speech was interrupted
+//       }
+//       // If we were just paused
+//       if (!_finished && !_playing) {
+//         try {
+//           // Wait to be unpaused
+//           await _sleeper.sleep();
+//         } catch (e) {
+//           // unpaused
+//         }
+//       }
+//     }
+//     await AudioServiceBackground.setState(
+//       controls: [],
+//       processingState: AudioProcessingState.stopped,
+//       playing: false,
+//     );
+//     if (!_finished) {
+//       onStop();
+//     }
+//     _completer.complete();
+//   }
+//
+//   @override
+//   Future<void> onPlay() => _playPause();
+//
+//   @override
+//   Future<void> onPause() => _playPause();
+//
+//   @override
+//   Future<void> onStop() async {
+//     // Signal the speech to stop
+//     _finished = true;
+//     _sleeper.interrupt();
+//     _tts.interrupt();
+//     // Wait for the speech to stop
+//     await _completer.future;
+//     // Shut down this task
+//     await super.onStop();
+//   }
+//
+//   MediaItem mediaItem(int number) => MediaItem(
+//       id: 'tts_$number',
+//       album: 'Numbers',
+//       title: 'Number $number',
+//       artist: 'Sample Artist');
+//
+//   Future<void> _playPause() async {
+//     if (_playing) {
+//       _interrupted = false;
+//       await AudioServiceBackground.setState(
+//         controls: [MediaControl.play, MediaControl.stop],
+//         processingState: AudioProcessingState.ready,
+//         playing: false,
+//       );
+//       _sleeper.interrupt();
+//       _tts.interrupt();
+//     } else {
+//       final session = await AudioSession.instance;
+//       // flutter_tts doesn't activate the session, so we do it here. This
+//       // allows the app to stop other apps from playing audio while we are
+//       // playing audio.
+//       if (await session.setActive(true)) {
+//         // If we successfully activated the session, set the state to playing
+//         // and resume playback.
+//         await AudioServiceBackground.setState(
+//           controls: [MediaControl.pause, MediaControl.stop],
+//           processingState: AudioProcessingState.ready,
+//           playing: true,
+//         );
+//         _sleeper.interrupt();
+//       }
+//     }
+//   }
+// }
 
 /// An object that performs interruptable sleep.
-class Sleeper {
-  Completer _blockingCompleter;
-
-  /// Sleep for a duration. If sleep is interrupted, a
-  /// [SleeperInterruptedException] will be thrown.
-  Future<void> sleep([Duration duration]) async {
-    _blockingCompleter = Completer();
-    if (duration != null) {
-      await Future.any([Future.delayed(duration), _blockingCompleter.future]);
-    } else {
-      await _blockingCompleter.future;
-    }
-    final interrupted = _blockingCompleter.isCompleted;
-    _blockingCompleter = null;
-    if (interrupted) {
-      throw SleeperInterruptedException();
-    }
-  }
-
-  /// Interrupt any sleep that's underway.
-  void interrupt() {
-    if (_blockingCompleter?.isCompleted == false) {
-      _blockingCompleter.complete();
-    }
-  }
-}
+// class Sleeper {
+//   Completer _blockingCompleter;
+//
+//   /// Sleep for a duration. If sleep is interrupted, a
+//   /// [SleeperInterruptedException] will be thrown.
+//   Future<void> sleep([Duration duration]) async {
+//     _blockingCompleter = Completer();
+//     if (duration != null) {
+//       await Future.any([Future.delayed(duration), _blockingCompleter.future]);
+//     } else {
+//       await _blockingCompleter.future;
+//     }
+//     final interrupted = _blockingCompleter.isCompleted;
+//     _blockingCompleter = null;
+//     if (interrupted) {
+//       throw SleeperInterruptedException();
+//     }
+//   }
+//
+//   /// Interrupt any sleep that's underway.
+//   void interrupt() {
+//     if (_blockingCompleter?.isCompleted == false) {
+//       _blockingCompleter.complete();
+//     }
+//   }
+// }
 
 class SleeperInterruptedException {}
 
 /// A wrapper around FlutterTts that makes it easier to wait for speech to
 /// complete.
-class Tts {
-  final FlutterTts _flutterTts = new FlutterTts();
-  Completer _speechCompleter;
-  bool _interruptRequested = false;
-  bool _playing = false;
-
-  Tts() {
-    _flutterTts.setCompletionHandler(() {
-      _speechCompleter?.complete();
-    });
-  }
-
-  bool get playing => _playing;
-
-  Future<void> speak(String text) async {
-    _playing = true;
-    if (!_interruptRequested) {
-      _speechCompleter = Completer();
-      await _flutterTts.speak(text);
-      await _speechCompleter.future;
-      _speechCompleter = null;
-    }
-    _playing = false;
-    if (_interruptRequested) {
-      _interruptRequested = false;
-      throw TtsInterruptedException();
-    }
-  }
-
-  Future<void> stop() async {
-    if (_playing) {
-      await _flutterTts.stop();
-      _speechCompleter?.complete();
-    }
-  }
-
-  void interrupt() {
-    if (_playing) {
-      _interruptRequested = true;
-      stop();
-    }
-  }
-}
+// class Tts {
+//   final FlutterTts _flutterTts = new FlutterTts();
+//   Completer _speechCompleter;
+//   bool _interruptRequested = false;
+//   bool _playing = false;
+//
+//   Tts() {
+//     _flutterTts.setCompletionHandler(() {
+//       _speechCompleter?.complete();
+//     });
+//   }
+//
+//   bool get playing => _playing;
+//
+//   Future<void> speak(String text) async {
+//     _playing = true;
+//     if (!_interruptRequested) {
+//       _speechCompleter = Completer();
+//       await _flutterTts.speak(text);
+//       await _speechCompleter.future;
+//       _speechCompleter = null;
+//     }
+//     _playing = false;
+//     if (_interruptRequested) {
+//       _interruptRequested = false;
+//       throw TtsInterruptedException();
+//     }
+//   }
+//
+//   Future<void> stop() async {
+//     if (_playing) {
+//       await _flutterTts.stop();
+//       _speechCompleter?.complete();
+//     }
+//   }
+//
+//   void interrupt() {
+//     if (_playing) {
+//       _interruptRequested = true;
+//       stop();
+//     }
+//   }
+// }
 
 class TtsInterruptedException {}
 
-class Seeker {
-  final AudioPlayer player;
-  final Duration positionInterval;
-  final Duration stepInterval;
-  final MediaItem mediaItem;
-  bool _running = false;
-
-  Seeker(
-    this.player,
-    this.positionInterval,
-    this.stepInterval,
-    this.mediaItem,
-  );
-
-  start() async {
-    _running = true;
-    while (_running) {
-      Duration newPosition = player.position + positionInterval;
-      if (newPosition < Duration.zero) newPosition = Duration.zero;
-      if (newPosition > mediaItem.duration) newPosition = mediaItem.duration;
-      player.seek(newPosition);
-      await Future.delayed(stepInterval);
-    }
-  }
-
-  stop() {
-    _running = false;
-  }
-}
+// class Seeker {
+//   final AudioPlayer player;
+//   final Duration positionInterval;
+//   final Duration stepInterval;
+//   final MediaItem mediaItem;
+//   bool _running = false;
+//
+//   Seeker(
+//     this.player,
+//     this.positionInterval,
+//     this.stepInterval,
+//     this.mediaItem,
+//   );
+//
+//   start() async {
+//     _running = true;
+//     while (_running) {
+//       Duration newPosition = player.position + positionInterval;
+//       if (newPosition < Duration.zero) newPosition = Duration.zero;
+//       if (newPosition > mediaItem.duration) newPosition = mediaItem.duration;
+//       player.seek(newPosition);
+//       await Future.delayed(stepInterval);
+//     }
+//   }
+//
+//   stop() {
+//     _running = false;
+//   }
+// }
